requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==3.0.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Cython==0.29.37
dill==0.3.8
distro==1.9.0
Django==4.1.13
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.5.0
docker==7.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.27.2
Faker==22.5.1
idna==3.6
inflection==0.5.1
isort==5.13.2
jsonschema==3.2.0
kombu==5.3.5
mccabe==0.7.0
mongoengine==0.27.0
packaging==23.2
paramiko==3.4.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==1.7.1
pylint==3.0.3
pylint-mongoengine==0.5.0
pylint-plugin-utils==0.8.2
pymongo==4.6.1
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
PyYAML==5.4.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
texttable==1.7.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.7.1
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websocket-client==0.59.0
